organize components and refactor them
