+++ client/src/components/videogame.jsx
@@ -1,14 +1,5 @@
 import React, { Component, Fragment } from 'react'
-import DevteamMember from './videogameDevteamMember'
-import PlatformBadge from './videogamePlatformBadge'
-import Tag from './videogameTag'
-import Suggested from './videogameSuggested'
 import Screen from './videogameScreen'
-import Review from './videogameReview'
-import ArchiveOffer from './videogameArchiveOffer'
-import Store from './videogameStore'
-import Company from './videogameCompany'
-import Genre from './videogameGenre'
 import VideogameSkeletonLoad from './videogameSkeletonLoad'
 import ImageGallery from './videogameImageGallery'
 import ImageGalleryCloseButton from './videogameImageGalleryCloseButton'
@@ -16,6 +7,10 @@ import Youtube from './videogameYoutube'
 import YoutubeGallery from './videogameYoutubeGallery'
 import YoutubeGalleryCloseButton from './videogameYoutubeGalleryCloseButton'
 import OneyGallery from './videogameOneyGallery'
+import Overview from './videogameOverview'
+import StoresAndReviews from './videogameStoresAndReviews'
+import SimilarVideogames from './videogameSimilarVideogames'
+import HeaderOnVideogames from './videogameHeader'
 
 class Videogame extends Component {
   state = {
@@ -24,7 +19,6 @@ class Videogame extends Component {
     archiveIdentifier: null,
     archiveOfferAvailable: false,
     id: this.props.selectedVideogame,
-    reviewHeight: '133px',
     galleryIsOpened: false,
     youtubeGalleryIsOpened: false,
     oneyGalleryIsOpened: false
@@ -71,46 +65,6 @@ class Videogame extends Component {
     }
   }
 
-  getReleaseYear = () => {
-    try {
-      return this.state.data.released ? this.state.data.released.match(/[0-9]{4}/) : 'n/a'
-    } catch (e) {
-      console.error(e)
-    }
-  }
-
-  getReleaseDate = () => {
-    try {
-      return this.state.data.released
-    } catch (e) {
-      console.error(e)
-    }
-  }
-
-  getOverview = () => {
-    try {
-      return this.state.data.description
-    } catch (e) {
-      console.error(e)
-    }
-  }
-
-  getPlaytime = () => {
-    try {
-      return this.state.data.playtime
-    } catch (e) {
-      console.error(e)
-    }
-  }
-
-  getVotes = () => {
-    try {
-      return this.state.data.rating
-    } catch (e) {
-      console.error(e)
-    }
-  }
-
   getClips = () => {
     try {
       let clip, poster
@@ -125,58 +79,6 @@ class Videogame extends Component {
     } catch (e) {
       console.error(e)
     }
-  }
-
-  getWebsite = () => {
-    try {
-      const website = this.state.data.website
-      const websiteText = website.replace(/http:\/\/|https:\/\/|www\./g, '')
-      return { website, websiteText }
-    } catch (e) {
-      console.error(e)
-    }
-  }
-
-  getBackground = () => {
-    try {
-      let background
-      this.state.data.background_image_additional
-        ? (background = this.state.data.background_image_additional)
-        : (background = this.state.data.background_image)
-
-      if (background) {
-        background.match(/media\/screenshots/)
-          ? (background = background.replace('media/screenshots', 'media/crop/600/400/screenshots'))
-          : (background = background.replace('media/games', 'media/crop/600/400/games'))
-      }
-      return background
-    } catch (e) {
-      console.error(e)
-    }
-  }
-
-  getPoster = () => {
-    try {
-      let poster
-      if (this.state.data.background_image) {
-        this.state.data.background_image.match(/media\/screenshots/)
-          ? (poster = this.state.data.background_image.replace('media/screenshots', 'media/crop/600/400/screenshots'))
-          : (poster = this.state.data.background_image.replace('media/games', 'media/crop/600/400/games'))
-      }
-      return poster
-    } catch (e) {
-      console.error(e)
-    }
-  }
-
-  unnecessaryReviewReadmore = () => {
-    let unnecessaryReviewReadmore
-    if (this.state.data.reviews[0]) {
-      this.state.data.reviews.length <= 1 && this.state.data.reviews[0].text.length < 445
-        ? (unnecessaryReviewReadmore = true)
-        : (unnecessaryReviewReadmore = false)
-    }
-    return unnecessaryReviewReadmore
   }
 
   setGalleryOpen = () => {
@@ -203,163 +105,20 @@ class Videogame extends Component {
     this.setState({ oneyGalleryIsOpened: false })
   }
 
-  setReviewsHeight = () => {
-    this.setState({ reviewHeight: 'auto' })
-  }
-
-  setBackReviewsHeight = () => {
-    this.setState({ reviewHeight: '133px' })
-  }
-
   render() {
     const data = this.state.data
-    const bgImage = this.state.dataIsReady
-      ? 'linear-gradient(rgba(0,0,0,.9), rgba(52,58,64,.9)), url(' + this.getBackground() + ')'
-      : 'url(data:image/gif;base64,R0lGODlhAQABAAAAACH5BAEKAAEALAAAAAABAAEAAAICTAEAOw==)'
 
     return (
       <Fragment>
         {this.state.dataIsReady ? (
           <main className='container'>
-            <header id='videogameHeader' border-bottom='1px' solid='#000'>
-              <h2 className='display-4 mt-2 heading-line' id='videogameLabel' display='inline'>
-                {this.getTitle()}
-                <span className='lead heading-line'> ({this.getReleaseYear()}) </span>
-              </h2>
-              {data.platforms.map((platformBadge, i) => (
-                <PlatformBadge key={i} data={platformBadge} />
-              ))}
-            </header>
-            {data.tags.length < 1 ? (
-              <section id='tags' className='my-2'>
-                {' '}
-              </section>
-            ) : (
-              <section id='tags' className='my-2'>
-                {data.tags.map((tag, i) => (
-                  <Tag key={i} data={tag} />
-                ))}
-              </section>
-            )}
-            <section
-              id='videogameSummary'
-              className='row text-white img-background details-background'
-              style={{ backgroundImage: bgImage }}>
-              <summary className='col-md-3 my-3'>
-                <img src={this.getPoster()} onClick={this.setGalleryOpen} alt='poster' className='img-style' />
-                <div className='my-3'>
-                  <h4>Facts:</h4>
-                  <ul className='list-unstyled'>
-                    <li>
-                      <strong>Company:</strong>{' '}
-                      {data.developers.map((companyElement, i) => (
-                        <Company key={companyElement.id} data={companyElement} index={i} />
-                      ))}
-                    </li>
-                    <li>
-                      <strong>Playtime:</strong>{' '}
-                      {this.getPlaytime() && this.getPlaytime() !== 0 ? this.getPlaytime() + ' hours' : '-'}
-                    </li>
-                    <li>
-                      <strong>Genre:</strong>{' '}
-                      {data.genres.map((genreElement, i) => (
-                        <Genre key={i} data={genreElement} />
-                      ))}
-                    </li>
-                    <li>
-                      <strong>Release:</strong> {this.getReleaseDate()}
-                    </li>
-                    <li>
-                      <strong>Voted:</strong> ★{this.getVotes()}/5
-                    </li>
-                    {this.getWebsite().website ? (
-                      <Fragment>
-                        <li>
-                          <strong>Web:</strong>{' '}
-                          <a
-                            href={this.getWebsite().website}
-                            target='_blank'
-                            rel='noopener noreferrer'
-                            className='text-secondary'>
-                            {this.getWebsite().websiteText.length >= 25
-                              ? this.getWebsite().websiteText.substring(0, 25) + '...'
-                              : this.getWebsite().websiteText}
-                          </a>
-                        </li>
-                      </Fragment>
-                    ) : null}
-                  </ul>
-                </div>
-              </summary>
-              <article id='overview' className='col my-3'>
-                <div>
-                  <h4>Overview:</h4>
-                  <p className='mb-2' dangerouslySetInnerHTML={{ __html: this.getOverview() }}></p>
-                </div>
-                <div>
-                  {data.devteam.length > 0 ? (
-                    <Fragment>
-                      <h4>Creators:</h4>
-                      <ul className='row list-unstyled list-group list-group-horizontal'>
-                        {data.devteam.map(devteamMember => (
-                          <DevteamMember key={devteamMember.id} data={devteamMember} />
-                        ))}
-                      </ul>
-                    </Fragment>
-                  ) : null}
-                </div>
-              </article>
-            </section>
-            <section id='storesAndReviews' className='row'>
-              <div className='col-md-3'>
-                {data.stores.length > 0 || this.state.archiveOfferAvailable ? (
-                  <Fragment>
-                    <div className='row mt-3 px-3'>
-                      <h4>Stores:</h4>
-                    </div>
-                    <div className='row mb-2'>
-                      <Fragment>
-                        {data.stores.length > 0
-                          ? data.stores.map(storeElement => <Store key={storeElement.id} data={storeElement} />)
-                          : null}
-                      </Fragment>
-                      {this.state.archiveOfferAvailable ? (
-                        <ArchiveOffer archiveIdentifier={this.state.archiveIdentifier} />
-                      ) : null}
-                    </div>
-                  </Fragment>
-                ) : null}
-              </div>
-              <article id='reviews' className='col'>
-                {data.reviews.length > 0 ? (
-                  <Fragment>
-                    <h4 className='row mt-3 px-3'>Reviews:</h4>
-                    <div id='longContent' className='long-content' style={{ height: this.state.reviewHeight }}>
-                      {data.reviews.map(reviewElement => (
-                        <Review key={reviewElement.id} data={reviewElement} />
-                      ))}
-                    </div>
-                    <div className='row justify-content-center'>
-                      {this.state.reviewHeight !== 'auto' ? (
-                        <Fragment>
-                          {!this.unnecessaryReviewReadmore() ? (
-                            <Fragment>
-                              <button className='btn btn-outline-dark text-center m-3' onClick={this.setReviewsHeight}>
-                                read more
-                              </button>{' '}
-                            </Fragment>
-                          ) : null}
-                        </Fragment>
-                      ) : (
-                        <button className='btn btn-outline-dark text-center m-3' onClick={this.setBackReviewsHeight}>
-                          read less
-                        </button>
-                      )}
-                    </div>
-                  </Fragment>
-                ) : null}
-              </article>
-            </section>
+            <HeaderOnVideogames data={data} />
+            <Overview data={data} dataIsReady={this.state.dataIsReady} />
+            <StoresAndReviews
+              data={data}
+              archiveIdentifier={this.state.archiveIdentifier}
+              archiveOfferAvailable={this.state.archiveOfferAvailable}
+            />
             {/* galleries start here */}
             <section id='outerImageRow' className='row'>
               <section id='imageGallery' className='col'>
@@ -472,17 +231,7 @@ class Videogame extends Component {
                 ) : null}
               </Fragment>
             </section>
-
-            <aside id='similarVideogames'>
-              <header className='row my-3 px-3'>
-                <h4>Similar titles:</h4>
-              </header>
-              <section className='row mb-2 justify-content-center text-center'>
-                {data.suggested.map(suggestedElement => (
-                  <Suggested key={suggestedElement.id} data={suggestedElement} />
-                ))}
-              </section>
-            </aside>
+            <SimilarVideogames data={data} />
           </main>
         ) : (
           <VideogameSkeletonLoad />
