+++ client/src/components/videogameYoutube.jsx
@@ -5,7 +5,7 @@ class Youtube extends Component {
     const youtubeElement = this.props.data.thumbnails.medium.url
     return (
       <div className='col-md-3 my-3'>
-        <span className='position-absolute bg-dark text-light youtube-play-style'>&#11208;</span>
+        <span className='position-absolute bg-dark text-light youtube-play-style'>►</span>
         {youtubeElement ? (
           <img className={`img-style ${this.props.class}`} src={youtubeElement} alt='screenshot of the video' />
         ) : null}
